feat: 검색 페이지 스타일 마이그레이션
